+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - MAJOR: big overhaul, all APIs changed, the first step is to go through client creation first, then you do not need to redefine the client options for each call
 - MAJOR: throughout the APIs, use dictionary parameters rather than multiple optional params
 - MAJOR: getEntries replaces getAllEntries, getEntriesByTypeId and getEntriesByIds
+- MAJOR: getAssetStreamById was renamed to getAssetById
 - MAJOR: validateSession was removed (incorrectly fell back to getApplicationStatus).
 - MAJOR: removed the method to update the uuid of an existing session
 - MAJOR: renamed the uuid optional parameter (when creating a client) to deviceId